+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `IServiceCollectionExtensions.AddOptions` with custom section name
+- `IConfigurationExtensions.GetValue` with strongly typed key
+
 ## [2.0.0] - 2026-05-05
 
 ### Changed