+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2026-09-14
+
 ### Added
 
 - `IServiceCollectionExtensions.AddOptions` with custom section name
@@ -34,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `IFormFile` extension method (AspNetCore package only)
 - `DbSet` extension methods for ISortableEntity interface (EntityFrameworkCore package only)
 
-[Unreleased]: https://github.com/neolution-ch/Neolution.Utilities/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/neolution-ch/Neolution.Utilities/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/neolution-ch/Neolution.Utilities/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/neolution-ch/Neolution.Utilities/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/neolution-ch/Neolution.Utilities/compare/v1.0.0-rc.1...v1.0.0